--- test/registered/function_call/test_qwen3_coder_detector/test_fixtures.py
@@ -1,59 +1,0 @@
-from sglang.srt.entrypoints.openai.protocol import Function, Tool
-from sglang.test.ci.ci_register import register_cpu_ci
-
-register_cpu_ci(1.0, "default")
-
-
-def get_test_tools():
-    return [
-        Tool(
-            type="function",
-            function=Function(
-                name="get_current_weather",
-                parameters={
-                    "type": "object",
-                    "properties": {
-                        "location": {"type": "string"},
-                        "unit": {"type": "string", "enum": ["celsius", "fahrenheit"]},
-                        "days": {"type": "integer"},
-                    },
-                    "required": ["location"],
-                },
-            ),
-        ),
-        Tool(
-            type="function",
-            function=Function(
-                name="sql_interpreter",
-                parameters={
-                    "type": "object",
-                    "properties": {
-                        "query": {"type": "string"},
-                        "dry_run": {"type": "boolean"},
-                    },
-                },
-            ),
-        ),
-        Tool(
-            type="function",
-            function=Function(
-                name="TodoWrite",
-                parameters={
-                    "type": "object",
-                    "properties": {
-                        "todos": {
-                            "type": "array",
-                            "items": {
-                                "type": "object",
-                                "properties": {
-                                    "content": {"type": "string"},
-                                    "status": {"type": "string"},
-                                },
-                                "required": ["content", "status"],
-                            },
-                        },
-                    },
-                },
-            ),
-        ),
-    ]
